Consider winsdk_samples a system library.
Since these headers are not part of WebRTC, this CL tells clang to
include them as system headers (in order to stop triggering warnings on
them).

This is part of the effort that will reduce the number of warnings
suppression flags propagated in the WebRTC build graph (by config and
public_configs).

Bug: webrtc:9251, webrtc:6269
Change-Id: I3879fe90627cca982f7b360e7bd8ca11592515bc
Reviewed-on: https://webrtc-review.googlesource.com/75514
Reviewed-by: Patrik Höglund <phoglund@webrtc.org>